Fix en_UK JavaScript message breakage
The message is used in JavaScript and ended up as

alert('Link was successfully copied to the clipboard. You can now use the system's paste functionality to share it!');

where the "system's" apostrophe breaks the quoting level, leading to no
image being displayed.

Use "system\\\'s" in source to get "system\'s" in JavaScript.